# klient/machine: improve indexing performance with concurrent file scanning