First step when creating a mount is to create the index from remote FS. For entire Koding repository we have following numbers:

**Scanning time**:  43.349127583s
**Files**:  212972
**Files time**:  2.838µs
**Files Disk Size**:  3.6 GiB
**Files Disk Size time**:  16.514726ms

The most time consuming operation is SHA-1 sum of stored files (which is expected and cached after the first scan). This operation can be made concurrently and that's what this PR introduces. Time measurements after the change:

**Scanning time**:  19.260137022s (improvement: **225%**)
**Files**:  212972
**Files time**:  8.605µs
**Files Disk Size**:  3.6 GiB
**Files Disk Size time**:  16.457452ms

## Motivation and Context
Decrease mount initialization time (which currently times out after one minute - due to Kite connection deadline).

## How Has This Been Tested?
Existing unit tests passed.

## Types of changes
- [x] New feature (non-breaking change which adds functionality)