@cihangir 2017-01-27
> most time consuming operation is SHA-1 sum

what about using another hashing algo while having 10x more performance 😛