@rjeczalik 2017-01-27
And in case we need parity with git index we can force index recreation

Cihangir <notifications@github.com> schrieb am Fr. 27. Jan. 2017 um 10:01:

> most time consuming operation is SHA-1 sum
>
> what about using another hashing algo while having 10x more performance 😛
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/koding/koding/pull/10434#issuecomment-275618486>, or mute
> the thread
> <https://github.com/notifications/unsubscribe-auth/ABG7ISUT8sKgT2S_QRhqenDREYJji3Ayks5rWbKGgaJpZM4LvbUr>
> .
>