@ppknap 2017-01-27
> what about using another hashing algo while having 10x more performance :stuck_out_tongue:

> And in case we need parity with git index we can force index recreation

OK, lets use crc32 as @cihangir sugessted [here](https://github.com/koding/koding/pull/10134#discussion_r94128683). We version the index type so it should be easy to switch anyway.