@ppknap 2017-01-27
> could you also send the code for measuring these?

@cihangir this is rather a dummy program than a real benchmark (I planning to add real bench when we have more time) but if you are interested, the code is [here](https://gist.github.com/ppknap/e4e9e68c46268555b5fb13d3e16b2f63)