@ppknap 2017-01-27
@cihangir - after switching to CRC-32 i got following results:
**SHA-1**:  13.191289437s
**CRC-32**:  11.610115897s

This was surprising so I profiled the index and most of the time spent is I/O during hashing. After using byte pools: 

**CRC-32 + BytePool**: Scanning time:  7.604643191s

I'm stopping at this point - need to focus on serialized file size optimization. Moreover, rest of the time is IO:
![image](https://cloud.githubusercontent.com/assets/5021246/22381945/ed453d90-e4c3-11e6-901b-ccd9399d81f7.png)